+++ Src/Public/Register-LabMedia.ps1
@@ -70,9 +70,14 @@ function Register-LabMedia {
         [ValidateSet('Windows','Linux')]
         [System.String] $OperatingSystem = 'Windows',
 
-        ## Registers media via a JSON file hosted externally
+        ## Registers media via a JSON file hosted externally.
         [Parameter(Mandatory, ValueFromPipelineByPropertyName, ParameterSetName = 'FromUri')]
         [System.String] $FromUri,
+
+        ## Registers media using a custom Id.
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'FromUri')]
+        [Parameter(ValueFromPipelineByPropertyName, ParameterSetName = 'Legacy')]
+        [System.String] $CustomId,
 
         ## Specifies that an exiting media entry should be overwritten.
         [Parameter(ValueFromPipelineByPropertyName)]
@@ -103,6 +108,11 @@ function Register-LabMedia {
                 ## Download the json content and convert into a hashtable
                 $legacyMedia = Get-LabMedia -Id $PSBoundParameters.Legacy -Legacy | Convert-PSObjectToHashtable;
 
+                if ($PSBoundParameters.ContainsKey('CustomId')) {
+
+                    $legacyMedia['Id'] = $CustomId
+                }
+
                 ## Recursively call Register-LabMedia and splat the properties
                 return (Register-LabMedia @legacyMedia -Force:$Force);
             }
@@ -111,6 +121,11 @@ function Register-LabMedia {
 
                 ## Download the json content and convert into a hashtable
                 $customMedia = Invoke-RestMethod -Uri $FromUri | Convert-PSObjectToHashtable;
+
+                if ($PSBoundParameters.ContainsKey('CustomId')) {
+
+                    $customMedia['Id'] = $CustomId
+                }
 
                 ## Recursively call Register-LabMedia and splat the properties
                 return (Register-LabMedia @customMedia -Force:$Force);
@@ -134,7 +149,15 @@ function Register-LabMedia {
                 }
 
                 ## Resolve the media Id to see if it's already been used
-                $media = Resolve-LabMedia -Id $Id -ErrorAction SilentlyContinue;
+                try {
+
+                    $media = Resolve-LabMedia -Id $Id -ErrorAction SilentlyContinue;
+                }
+                catch {
+
+                    Write-Debug -Message ($localized.CannotLocateMediaError -f $Id);
+                }
+
                 if ($media -and (-not $Force)) {
 
                     throw ($localized.MediaAlreadyRegisteredError -f $Id, '-Force');